♻️ refactor: home SSG로 변경
