+++ src/libs/apis/project.ts
@@ -13,13 +13,9 @@ export const createProject = async (data: ProjectDataResponse) => {
 }
 
 // 전체 프로젝트 조회 (READ)
-export const getAllProjects = async (accessToken: string) => {
+export const getAllProjects = async () => {
   try {
-    const response = await axiosInstance.get('/projects/all', {
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-      },
-    })
+    const response = await axiosInstance.get('/projects/all')
     return response.data
   } catch (error) {
     console.error('프로젝트 조회 실패:', error)
